+++ src/extension.ts
@@ -3,7 +3,7 @@ import * as vscode from 'vscode';
 import * as nls from 'vscode-nls';
 import TelemetryReporter from 'vscode-extension-telemetry';
 import { BrowserPreview } from './editorPreview/browserPreview';
-import { Manager } from './manager';
+import { ServerGrouping } from './serverGrouping';
 import { EXTENSION_ID } from './utils/constants';
 import { PathUtil } from './utils/pathUtil';
 import {
@@ -11,9 +11,11 @@ import {
 	SETTINGS_SECTION_ID,
 	SettingUtil,
 } from './utils/settingsUtil';
+import { existsSync } from 'fs';
 
 let reporter: TelemetryReporter;
-let manager: Manager;
+let managers: Map<vscode.Uri, ServerGrouping>;
+let looseFileManager: ServerGrouping;
 
 const localize = nls.loadMessageBundle();
 
@@ -25,12 +27,23 @@ export function activate(context: vscode.ExtensionContext): void {
 		extPackageJSON.aiKey
 	);
 
-	manager = new Manager(
-		context.extensionUri,
-		reporter,
-		PathUtil.GetUserDataDirFromStorageUri(context.storageUri?.fsPath)
-	);
+		managers = new Map<vscode.Uri, ServerGrouping>();
+	// let manager = new HostedContent(
+	// 	context.extensionUri,
+	// 	reporter,
+	// 	PathUtil.GetUserDataDirFromStorageUri(context.storageUri?.fsPath)
+	// );
 
+	const createHostedContentForWorkspace = (
+		workspace: vscode.WorkspaceFolder | undefined
+	) => {
+		return new ServerGrouping(
+				context.extensionUri,
+				reporter,
+				workspace,
+				PathUtil.GetUserDataDirFromStorageUri(context.storageUri?.fsPath)
+			);
+	};
 	/* __GDPR__
 		"extension.startUp" : {
 			"numWorkspaceFolders" : { "classification": "SystemMetaData", "purpose": "FeatureInsight", "isMeasurement": true }
@@ -45,43 +58,146 @@ export function activate(context: vscode.ExtensionContext): void {
 	const openPreview = (
 		internal: boolean,
 		file: string,
+		hc: ServerGrouping,
 		isRelative: boolean,
-		debug = false
+		debug = false,
 	) => {
 		if (internal) {
 			// for now, ignore debug or no debug for embedded preview
-			manager.createOrShowEmbeddedPreview(undefined, file, isRelative);
+			hc.createOrShowEmbeddedPreview(undefined, file, isRelative);
 		} else {
-			manager.showPreviewInBrowser(file, isRelative, debug);
+			hc.showPreviewInBrowser(file, isRelative, debug);
 		}
 	};
 
+	const getHCFromWorkspace = (workspace: vscode.WorkspaceFolder) => {
+		const hcFromMap = managers.get(workspace.uri);
+					if (hcFromMap) {
+						return hcFromMap;
+					} else {
+						const hc = createHostedContentForWorkspace(workspace);
+						managers.set(workspace.uri, hc);
+						return hc;
+					}
+	};
+
+	const getLooseFileHC = () => {
+		if (!looseFileManager) {
+			looseFileManager = createHostedContentForWorkspace(undefined);
+		}
+		return looseFileManager;
+	};
+
+	const determineHCfromFile = (file: vscode.Uri | string, fileStringRelative: boolean) => {
+		if (fileStringRelative) {
+			getLooseFileHC();
+		} else {
+			let fileUri;
+			if (typeof file == 'string') {
+				fileUri = vscode.Uri.parse(file);
+			} else if (file instanceof vscode.Uri) {
+				fileUri = file;
+			} else {
+
+			getLooseFileHC();
+			}
+			if (fileUri) {
+				const workspace = vscode.workspace.getWorkspaceFolder(fileUri);
+				if (workspace) {
+					return getHCFromWorkspace(workspace);
+				}
+			}
+		}
+	};
+
+	const handleOpenFileCaller = (
+		internal: boolean,
+		file: vscode.Uri | string | undefined,
+		fileStringRelative = true,
+		debug = false,
+		workspace?: vscode.WorkspaceFolder,
+		port?: number,
+		hc?: ServerGrouping
+		) => {
+			if (!file) {
+				openNoTarget();
+				return;
+			}
+			if (!hc) {
+				if (workspace) {
+					hc = getHCFromWorkspace(workspace);
+				} else if (port) {
+					managers.forEach((potentialHC,key) => {
+						if (potentialHC.port === port) {
+							hc = potentialHC;
+							return;
+						}
+						hc = determineHCfromFile(file, fileStringRelative);
+					});
+				} else {
+					hc = determineHCfromFile(file, fileStringRelative);
+				}
+			}
+			if (hc) {
+				handleOpenFile(internal,file,hc,fileStringRelative,debug);
+			}
+
+	};
+
+	const openNoTarget = () => {
+		const workspaces = vscode.workspace.workspaceFolders;
+				if (workspaces && workspaces.length > 0) {
+					for (let i = 0; i < workspaces.length; i++) {
+						const currWorkspace = workspaces[i];
+						const manager = managers.get(currWorkspace.uri);
+						if (manager) {
+							vscode.commands.executeCommand(
+								`${SETTINGS_SECTION_ID}.start.preview.atFile`,
+								'/',
+								true,
+								currWorkspace,
+								undefined,
+								manager
+							);
+							return;
+						}
+					}
+				} else {
+					vscode.commands.executeCommand(
+						`${SETTINGS_SECTION_ID}.start.preview.atFile`,
+						'/',
+						false
+					);
+				}
+	};
 	const handleOpenFile = (
 		internal: boolean,
-		file: any,
+		file: vscode.Uri | string,
+		hc: ServerGrouping,
 		fileStringRelative = true,
 		debug = false
 	) => {
 		if (typeof file == 'string') {
-			openPreview(internal, file, fileStringRelative, debug);
+			openPreview(internal, file, hc, fileStringRelative, debug);
 			return;
 		} else if (file instanceof vscode.Uri) {
+			console.log(vscode.workspace.getWorkspaceFolder(file));
 			const filePath = file?.fsPath;
 			if (filePath) {
-				openPreview(internal, filePath, false, debug);
+				openPreview(internal, filePath, hc, false, debug);
 				return;
 			} else {
 				const activeFilePath =
 					vscode.window.activeTextEditor?.document.fileName;
 				if (activeFilePath) {
-					openPreview(internal, activeFilePath, false, debug);
+					openPreview(internal, activeFilePath, hc,false, debug);
 					return;
 				}
 			}
 		} else {
 			const activeFilePath = vscode.window.activeTextEditor?.document.fileName;
 			if (activeFilePath) {
-				openPreview(internal, activeFilePath, false, debug);
+				openPreview(internal, activeFilePath, hc,false, debug);
 				return;
 			}
 		}
@@ -103,21 +219,31 @@ export function activate(context: vscode.ExtensionContext): void {
 				context.extensionUri
 			).defaultPreviewPath;
 			if (filePath == '') {
-				if (manager.workspace) {
+				openNoTarget();
+
+			} else {
+				managers.forEach((manager)=> {
+					if (manager.pathExistsRelativeToWorkspace(filePath)) {
+						vscode.commands.executeCommand(
+							`${SETTINGS_SECTION_ID}.start.preview.atFile`,
+							filePath,
+							true,
+							manager.workspace,
+							undefined,
+							manager
+						);
+						return;
+					}
+				});
+				if (existsSync(filePath)) {
 					vscode.commands.executeCommand(
 						`${SETTINGS_SECTION_ID}.start.preview.atFile`,
-						'/',
-						true
+						filePath,
+						false
 					);
 				} else {
-					manager.openServer();
+					throw Error();
 				}
-			} else {
-				vscode.commands.executeCommand(
-					`${SETTINGS_SECTION_ID}.start.preview.atFile`,
-					filePath,
-					manager.pathExistsRelativeToWorkspace(filePath)
-				);
 			}
 		})
 	);
@@ -125,12 +251,15 @@ export function activate(context: vscode.ExtensionContext): void {
 	context.subscriptions.push(
 		vscode.commands.registerCommand(
 			`${SETTINGS_SECTION_ID}.start.preview.atFile`,
-			(file?: any, relativeFileString = true) => {
+			(file?: vscode.Uri | string, options?:any, relativeFileString = false, workspace?: vscode.WorkspaceFolder, port?: number, manager?: ServerGrouping) => {
 				const previewType = SettingUtil.GetPreviewType(context.extensionUri);
 				vscode.commands.executeCommand(
 					`${SETTINGS_SECTION_ID}.start.${previewType}.atFile`,
 					file,
-					relativeFileString
+					relativeFileString,
+					workspace,
+					port,
+					manager
 				);
 			}
 		)
@@ -139,7 +268,7 @@ export function activate(context: vscode.ExtensionContext): void {
 	context.subscriptions.push(
 		vscode.commands.registerCommand(
 			`${SETTINGS_SECTION_ID}.start.debugPreview.atFile`,
-			(file?: any, relativeFileString = true) => {
+			(file?: vscode.Uri | string, relativeFileString = true, workspace?: vscode.WorkspaceFolder, port?: number, manager?: ServerGrouping) => {
 				// TODO: implement internalDebugPreview and use settings to choose which one to launch
 				vscode.commands.executeCommand(
 					`${SETTINGS_SECTION_ID}.start.externalDebugPreview.atFile`,
@@ -153,7 +282,7 @@ export function activate(context: vscode.ExtensionContext): void {
 	context.subscriptions.push(
 		vscode.commands.registerCommand(
 			`${SETTINGS_SECTION_ID}.start.externalPreview.atFile`,
-			(file?: any, relativeFileString = true) => {
+			(file?: vscode.Uri | string, relativeFileString = false, workspace?: vscode.WorkspaceFolder, port?: number, manager?: ServerGrouping) => {
 				/* __GDPR__
 					"preview" :{
 						"type" : {"classification": "SystemMetaData", "purpose": "FeatureInsight"},
@@ -165,7 +294,7 @@ export function activate(context: vscode.ExtensionContext): void {
 					location: 'atFile',
 					debug: 'false',
 				});
-				handleOpenFile(false, file, relativeFileString);
+				handleOpenFileCaller(false, file, relativeFileString, false, workspace,port, manager);
 			}
 		)
 	);
@@ -173,7 +302,7 @@ export function activate(context: vscode.ExtensionContext): void {
 	context.subscriptions.push(
 		vscode.commands.registerCommand(
 			`${SETTINGS_SECTION_ID}.start.internalPreview.atFile`,
-			(file?: any, relativeFileString = true) => {
+			(file?: vscode.Uri | string, relativeFileString = false, workspace?: vscode.WorkspaceFolder, port?: number, manager?: ServerGrouping) => {
 				/* __GDPR__
 					"preview" :{
 						"type" : {"classification": "SystemMetaData", "purpose": "FeatureInsight"},
@@ -184,7 +313,7 @@ export function activate(context: vscode.ExtensionContext): void {
 					type: 'internal',
 					location: 'atFile',
 				});
-				handleOpenFile(true, file, relativeFileString);
+				handleOpenFileCaller(true, file, relativeFileString, false, workspace,port, manager);
 			}
 		)
 	);
@@ -192,7 +321,7 @@ export function activate(context: vscode.ExtensionContext): void {
 	context.subscriptions.push(
 		vscode.commands.registerCommand(
 			`${SETTINGS_SECTION_ID}.start.externalDebugPreview.atFile`,
-			(file?: any, relativeFileString = true) => {
+			(file?: vscode.Uri | string, relativeFileString = false, workspace?: vscode.WorkspaceFolder, port?: number, manager?: ServerGrouping) => {
 				/* __GDPR__
 					"preview" :{
 						"type" : {"classification": "SystemMetaData", "purpose": "FeatureInsight"},
@@ -204,20 +333,23 @@ export function activate(context: vscode.ExtensionContext): void {
 					location: 'atFile',
 					debug: 'true',
 				});
-				handleOpenFile(false, file, relativeFileString, true);
+
+
+				handleOpenFileCaller(false, file, relativeFileString, true, workspace, port, manager);
 			}
 		)
 	);
 
 	context.subscriptions.push(
 		vscode.commands.registerCommand(`${SETTINGS_SECTION_ID}.end`, () => {
-			if (!manager.closeServer()) {
-				/* __GDPR__
-					"server.forceClose" : {}
-				*/
-				reporter.sendTelemetryEvent('server.forceClose');
-				vscode.window.showErrorMessage('Server already off.');
-			}
+			/* __GDPR__
+				"server.forceClose" : {}
+			*/
+			managers.forEach((manager: ServerGrouping, key: vscode.Uri) => {
+				manager.closeServer();
+			});
+			reporter.sendTelemetryEvent('server.forceClose');
+
 		})
 	);
 
@@ -225,53 +357,68 @@ export function activate(context: vscode.ExtensionContext): void {
 		vscode.commands.registerCommand(
 			`${SETTINGS_SECTION_ID}.setDefaultOpenFile`,
 			(file: vscode.Uri) => {
-				if (manager.absPathInDefaultWorkspace(file.fsPath)) {
-					const fileRelativeToWorkspace =
-						manager.getFileRelativeToDefaultWorkspace(file.fsPath) ?? '';
-					SettingUtil.UpdateSettings(
-						Settings.defaultPreviewPath,
-						fileRelativeToWorkspace,
-						false
-					);
-				} else {
-					SettingUtil.UpdateSettings(
-						Settings.defaultPreviewPath,
-						file.fsPath,
-						false
-					);
+
+				const workspace = vscode.workspace.getWorkspaceFolder(file);
+				if (workspace) {
+					const hc = new ServerGrouping(
+						context.extensionUri,
+						reporter,
+						workspace,
+						PathUtil.GetUserDataDirFromStorageUri(context.storageUri?.fsPath));
+					managers.set(workspace.uri,hc);
+
+					if (hc.absPathInDefaultWorkspace(file.fsPath)) {
+						const fileRelativeToWorkspace =
+						hc.getFileRelativeToDefaultWorkspace(file.fsPath) ?? '';
+						SettingUtil.UpdateSettings(
+							Settings.defaultPreviewPath,
+							fileRelativeToWorkspace,
+							false
+						);
+					} else {
+						SettingUtil.UpdateSettings(
+							Settings.defaultPreviewPath,
+							file.fsPath,
+							false
+						);
+					}
 				}
+
+
 			}
 		)
 	);
 
-	if (vscode.window.registerWebviewPanelSerializer) {
-		vscode.window.registerWebviewPanelSerializer(BrowserPreview.viewType, {
-			async deserializeWebviewPanel(
-				webviewPanel: vscode.WebviewPanel,
-				state: any
-			) {
-				let relative = true;
-				let file = state.currentAddress ?? '/';
+	// if (vscode.window.registerWebviewPanelSerializer) {
+	// 	vscode.window.registerWebviewPanelSerializer(BrowserPreview.viewType, {
+	// 		async deserializeWebviewPanel(
+	// 			webviewPanel: vscode.WebviewPanel,
+	// 			state: any
+	// 		) {
+	// 			let relative = true;
+	// 			let file = state.currentAddress ?? '/';
 
-				if (!manager.pathExistsRelativeToWorkspace(file)) {
-					const absFile = manager.decodeEndpoint(file);
-					file = absFile ?? '/';
-					relative = false;
-				}
+	// 			if (!manager.pathExistsRelativeToWorkspace(file)) {
+	// 				const absFile = manager.decodeEndpoint(file);
+	// 				file = absFile ?? '/';
+	// 				relative = false;
+	// 			}
 
-				if (file == '/' && !manager.workspace) {
-					// root will not show anything, so cannot revive content. Dispose.
-					webviewPanel.dispose();
-					return;
-				}
-				webviewPanel.webview.options = manager.getWebviewOptions();
-				manager.createOrShowEmbeddedPreview(webviewPanel, file, relative);
-			},
-		});
-	}
+	// 			if (file == '/' && !manager.workspace) {
+	// 				// root will not show anything, so cannot revive content. Dispose.
+	// 				webviewPanel.dispose();
+	// 				return;
+	// 			}
+	// 			webviewPanel.webview.options = manager.getWebviewOptions();
+	// 			manager.createOrShowEmbeddedPreview(webviewPanel, file, relative);
+	// 		},
+	// 	});
+	// }
 }
 
 export function deactivate(): void {
 	reporter.dispose();
-	manager.dispose();
+	managers.forEach((manager) => {
+		manager.dispose();
+	});
 }
