+++ src/task/serverTaskProvider.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import TelemetryReporter from 'vscode-extension-telemetry';
-import { serverMsg } from '../manager';
+import { serverMsg } from '../serverGrouping';
 import { EndpointManager } from '../infoManagers/endpointManager';
 import { WorkspaceManager } from '../infoManagers/workspaceManager';
 import { Disposable } from '../utils/dispose';
